Remove Hackday from playbook
We're thinking in other ways to invest our free time in a productive way. We want to use this time for stuff that help us learn new things, test new technologies, write blogposts/talks and help open source software.

The current hackday's format did no work as we expected:
* For most of the projects one day is not enough to deliver something meaninful
* One day per month is not productive
* Most of the time we did not organize the projects

We still see value in using the free time for investing in those stuff, but we would like to think in new formats that help us deliver real value.

Please comment in this PR so we can think in ways to overcome the problems.